move FuncTranslator into translator.rs
Also move VisitOperator for FuncBuilder into mod.rs.
