--- crates/wasmi/src/engine/func_builder/visit.rs
@@ -1,78 +1,0 @@
-use super::{FuncBuilder, FuncValidator, TranslationError};
-use wasmparser::{BinaryReaderError, VisitOperator};
-
-impl<'parser> FuncBuilder<'parser> {
-    /// Translates into `wasmi` bytecode if the current code path is reachable.
-    fn validate_then_translate<V, F>(
-        &mut self,
-        validate: V,
-        translator: F,
-    ) -> Result<(), TranslationError>
-    where
-        V: FnOnce(&mut FuncValidator) -> Result<(), BinaryReaderError>,
-        F: FnOnce(&mut Self) -> Result<(), TranslationError>,
-    {
-        validate(&mut self.validator)?;
-        translator(self)?;
-        Ok(())
-    }
-}
-
-macro_rules! impl_visit_operator {
-    ( @mvp BrTable { $arg:ident: $argty:ty } => $visit:ident $($rest:tt)* ) => {
-        // We need to special case the `BrTable` operand since its
-        // arguments (a.k.a. `BrTable<'a>`) are not `Copy` which all
-        // the other impls make use of.
-        fn $visit(&mut self, $arg: $argty) -> Self::Output {
-            let offset = self.current_pos();
-            let arg_cloned = $arg.clone();
-            self.validate_then_translate(
-                |v| v.visitor(offset).$visit(arg_cloned),
-                |this| this.$visit($arg),
-            )
-        }
-        impl_visit_operator!($($rest)*);
-    };
-    ( @mvp $($rest:tt)* ) => {
-        impl_visit_operator!(@@supported $($rest)*);
-    };
-    ( @sign_extension $($rest:tt)* ) => {
-        impl_visit_operator!(@@supported $($rest)*);
-    };
-    ( @saturating_float_to_int $($rest:tt)* ) => {
-        impl_visit_operator!(@@supported $($rest)*);
-    };
-    ( @bulk_memory $($rest:tt)* ) => {
-        impl_visit_operator!(@@supported $($rest)*);
-    };
-    ( @reference_types $($rest:tt)* ) => {
-        impl_visit_operator!(@@supported $($rest)*);
-    };
-    ( @@supported $op:ident $({ $($arg:ident: $argty:ty),* })? => $visit:ident $($rest:tt)* ) => {
-        fn $visit(&mut self $($(,$arg: $argty)*)?) -> Self::Output {
-            let offset = self.current_pos();
-            self.validate_then_translate(
-                |v| v.visitor(offset).$visit($($($arg),*)?),
-                |this| {
-                    this.$visit($($($arg),*)?)
-                },
-            )
-        }
-        impl_visit_operator!($($rest)*);
-    };
-    ( @$proposal:ident $op:ident $({ $($arg:ident: $argty:ty),* })? => $visit:ident $($rest:tt)* ) => {
-        // Wildcard match arm for all the other (yet) unsupported Wasm proposals.
-        fn $visit(&mut self $($(, $arg: $argty)*)?) -> Self::Output {
-            let offset = self.current_pos();
-            self.validator.visitor(offset).$visit($($($arg),*)?).map_err(::core::convert::Into::into)
-        }
-        impl_visit_operator!($($rest)*);
-    };
-    () => {};
-}
-
-impl<'a> VisitOperator<'a> for FuncBuilder<'a> {
-    type Output = Result<(), TranslationError>;
-
-    wasmparser::for_each_operator!(impl_visit_operator);
-}
